unique slug

+++ database/migrations/2023_02_01_044628_create_products_table.php
@@ -12,9 +12,15 @@ return new class extends Migration {
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('title');
-            $table->string('slug');
-            $table->string('thumbnail')->nullable();
-            $table->unsignedBigInteger('price')->default(0);
+
+            $table->string('slug')
+                ->unique();
+
+            $table->string('thumbnail')
+                ->nullable();
+
+            $table->unsignedBigInteger('price')
+                ->default(0);
 
             $table->timestamps();
 
@@ -28,6 +34,8 @@ return new class extends Migration {
 
     public function down(): void
     {
-        Schema::dropIfExists('products');
+        if (app()->isLocal()) {
+            Schema::dropIfExists('products');
+        }
     }
 };
